pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yolo-toys"
version = "3.1.0"
description = "多模型实时视觉识别系统 - 支持 YOLO、HuggingFace Transformers、多模态模型"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "LessUp", email = "lessup@users.noreply.github.com"}
]
keywords = [
    "yolo",
    "object-detection",
    "computer-vision",
    "deep-learning",
    "fastapi",
    "real-time",
    "segmentation",
    "pose-estimation",
    "huggingface",
    "transformers",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Typing :: Typed",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "ultralytics>=8.3.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.44.0",
    "Pillow>=10.4.0",
    "cachetools>=5.5.0",
    "prometheus-client>=0.21.0",
    "python-magic>=0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
    "pre-commit>=4.0.0",
    "pytest-cov>=6.0.0",
    "pip-audit>=2.7.0",
    "pytest-benchmark>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/LessUp/yolo-toys"
Repository = "https://github.com/LessUp/yolo-toys"
Documentation = "https://lessup.github.io/yolo-toys/"
Issues = "https://github.com/LessUp/yolo-toys/issues"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "C4"]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function call in argument defaults (FastAPI File/Query pattern)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
branch = true
omit = [
    "app/__init__.py",
    "app/main.py",  # Entry point, hard to test
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 70  # Minimum coverage threshold
